Converted Document to an abstract class and moved mutation to DocumentEditor.
